requirements.txt
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy>=1.24,<2
scikit-learn>=1.3
pandas>=2.0
rich>=13.0
pdfminer.six>=20221105
fastapi>=0.110
uvicorn[standard]>=0.29
pytest>=8.0
pytest-asyncio>=0.23
streamlit>=1.32.0
